moves utilities to payload
