+++ packages/next/src/views/Document/getDocumentData.tsx
@@ -6,8 +6,8 @@ import type {
   SanitizedGlobalConfig,
 } from 'payload'
 
-import { reduceFieldsToValues } from '@payloadcms/ui/shared'
 import { buildFormState } from '@payloadcms/ui/utilities/buildFormState'
+import { reduceFieldsToValues } from 'payload'
 
 export const getDocumentData = async (args: {
   collectionConfig?: SanitizedCollectionConfig
