+++ packages/payload/src/utilities/getDataByPath.ts
@@ -1,6 +1,6 @@
 import type { FormState } from 'payload'
 
-import { unflatten } from '../../utilities/unflatten.js'
+import { unflatten } from './unflatten.js'
 
 export const getDataByPath = <T = unknown>(fields: FormState, path: string): T => {
   const pathPrefixToRemove = path.substring(0, path.lastIndexOf('.') + 1)
