+++ packages/payload/src/utilities/reduceFieldsToValues.ts
