+++ packages/richtext-slate/src/field/elements/link/Button/index.tsx
@@ -3,7 +3,8 @@
 import type { FormState } from 'payload'
 
 import { useConfig, useDrawerSlug, useFieldProps, useModal, useTranslation } from '@payloadcms/ui'
-import { getFormState, reduceFieldsToValues } from '@payloadcms/ui/shared'
+import { getFormState } from '@payloadcms/ui/shared'
+import { reduceFieldsToValues } from 'payload'
 import React, { Fragment, useState } from 'react'
 import { Editor, Range, Transforms } from 'slate'
 import { ReactEditor, useSlate } from 'slate-react'
